refactor: Modify UpdateApplyStatusOptions and application form props
