+++ app/components/modal/club/applicationFormViewModal.tsx
@@ -1,4 +1,4 @@
-import React, { useEffect, useRef } from "react";
+import React, { useEffect } from "react";
 import Image from "next/image";
 import ResultBadge from "@/components/badge/resultBadge";
 import close from "@/images/icon/close.svg";
@@ -16,8 +16,6 @@ import UpdateApplyStatusOptions from "@/components/dropdown/updateApplyStatusOpt
 export interface ApplicationFormViewModalProps {
   applyId: string;
   onClose: () => void;
-  isOptionsOpen: boolean;
-  setIsOptionsOpen: (isOpen: boolean) => void;
   setSelectedOption: (option: string) => void;
   setIsModalOpen: (isOpen: boolean) => void;
 }
@@ -25,13 +23,9 @@ export interface ApplicationFormViewModalProps {
 const ApplicationFormViewModal = ({
   applyId,
   onClose,
-  isOptionsOpen,
-  setIsOptionsOpen,
   setSelectedOption,
   setIsModalOpen,
 }: ApplicationFromViewModalProps) => {
-  const optionsRef = useRef<HTMLDivElement | null>(null);
-
   const { applyDetail, isError, isLoading } = useApplyDetailQuery(applyId);
   const {
     applyData,
@@ -40,10 +34,6 @@ const ApplicationFormViewModal = ({
     fileUri,
     portfolioUrl,
   } = applyDetail ?? {};
-
-  const handleMenuClick = (label: string) => {
-    setSelectedOption(label);
-  };
 
   const handleClose = () => {
     onClose();
@@ -99,11 +89,8 @@ const ApplicationFormViewModal = ({
         <div className="flex gap-2">
           <UpdateApplyStatusOptions
             checkedApplications={[applyId]}
-            isOptionsOpen={isOptionsOpen}
-            setIsOptionsOpen={setIsOptionsOpen}
             setSelectedStatus={setSelectedOption}
             setIsModalOpen={setIsModalOpen}
-            optionsRef={optionsRef}
           />
           <ResultBadge status={APPLY_STATUS_MAP[applyData.applyStatusType]} />
         </div>
