+++ app/hooks/apply/useApplyStatusOptions.tsx
@@ -1,4 +1,4 @@
-import { useEffect, useRef, useState } from "react";
+import { useState } from "react";
 
 export const useApplyStatusOptions = () => {
   const [isOptionsOpen, setIsOptionsOpen] = useState<boolean>(false); // 지원상태 변경 옵션 드롭다운/바텀시트
@@ -7,27 +7,10 @@ export const useApplyStatusOptions = () => {
   const [isInterviewOpen, setIsInterviewOpen] = useState<boolean>(false); // 면접 확인 메세지 전송 모달
   const [isApplicationOpen, setIsApplicationOpen] = useState<boolean>(false); // 지원서 상세보기 모달
 
-  const optionsRef = useRef<HTMLDivElement>(null);
-
   const handleOptionSelect = (option: string) => {
     setSelectedOption(option);
     setIsModalOpen(true);
   };
-
-  useEffect(() => {
-    const handleClickOutside = (event: MouseEvent) => {
-      if (
-        optionsRef.current &&
-        !optionsRef.current.contains(event.target as Node)
-      ) {
-        setIsOptionsOpen(false);
-      }
-    };
-    document.addEventListener("mousedown", handleClickOutside);
-    return () => {
-      document.removeEventListener("mousedown", handleClickOutside);
-    };
-  }, []);
 
   return {
     isOptionsOpen,
@@ -41,6 +24,5 @@ export const useApplyStatusOptions = () => {
     setIsInterviewOpen,
     isApplicationOpen,
     setIsApplicationOpen,
-    optionsRef,
   };
 };
